Move Firefox specific code out of iteration #1065.
Lets move the geckoProfiler fixes to Firefox. First move the code
(and add new delegate hook) then in a follow up we can go through
and rename the hooks so they are easier to understand.
