+++ lib/chrome/webdriver/chromeDelegate.js
@@ -416,6 +416,8 @@ class ChromeDelegate {
     this.hars.push(har);
   }
 
+  async afterCollect() {}
+
   async sendAndGetDevToolsCommand(cmd, params = {}) {
     return this.cdpClient.send(cmd, params);
   }
