+++ lib/core/engine/iteration.js
@@ -4,7 +4,6 @@ const webdriver = require('selenium-webdriver');
 const log = require('intel').getLogger('browsertime');
 const SeleniumRunner = require('../seleniumRunner');
 const preURL = require('../../support/preURL');
-const pathToFolder = require('../../support/pathToFolder');
 const setResourceTimingBufferSize = require('../../support/setResourceTimingBufferSize');
 const ScreenshotManager = require('../../screenshot/');
 const engineUtils = require('../../support/engineUtils');
@@ -24,7 +23,6 @@ const Android = require('./command/android.js');
 const ChromeDevToolsProtocol = require('./command/chromeDevToolsProtocol.js');
 const { addConnectivity, removeConnectivity } = require('../../connectivity');
 const util = require('../../support/util.js');
-const path = require('path');
 
 const { isAndroidConfigured } = require('../../android');
 
@@ -227,63 +225,8 @@ class Iteration {
                 );
               }
             }
+            result[i].videoRecordingStart = videoMetrics.videoRecordingStart;
             result[i].visualMetrics = videoMetrics.visualMetrics;
-
-            if (
-              options.browser === 'firefox' &&
-              options.firefox &&
-              options.firefox.geckoProfiler
-            ) {
-              const profileFilename = `geckoProfile-${index}.json.gz`;
-              const profileSubdir = pathToFolder(result[i].url, options);
-
-              try {
-                const geckoProfile = JSON.parse(
-                  await this.storageManager.readData(
-                    profileFilename,
-                    profileSubdir
-                  )
-                );
-                if (!geckoProfile.meta) {
-                  geckoProfile.meta = {};
-                }
-
-                // Here we are calculating the unix timestamp of the first frame after orange screen
-                // by adding the time to first frame, time to orange screen to the unix timestamp of
-                // video recording start time
-                const firstFrameStartTime =
-                  result[i].recordingStartTime +
-                  videoMetrics.videoRecordingStart +
-                  result[i].timeToFirstFrame;
-
-                for (let progress of [
-                  'VisualProgress',
-                  'ContentfulSpeedIndexProgress',
-                  'PerceptualSpeedIndexProgress'
-                ]) {
-                  videoMetrics.visualMetrics[
-                    progress
-                  ] = util.adjustVisualProgressTimestamps(
-                    videoMetrics.visualMetrics[progress],
-                    geckoProfile.meta.startTime,
-                    firstFrameStartTime
-                  );
-                }
-                geckoProfile.meta.visualMetrics = videoMetrics.visualMetrics;
-
-                await this.storageManager.writeJson(
-                  path.join(profileSubdir, `geckoProfile-${index}.json`),
-                  geckoProfile,
-                  true
-                );
-              } catch (e) {
-                log.error(
-                  `Could not write visual metrics to ${profileFilename}`,
-                  e
-                );
-              }
-            }
-
             i++;
           } catch (e) {
             // If one of the Visual Metrics runs fail, just swallow and try the next one
@@ -291,6 +234,9 @@ class Iteration {
           }
         }
       }
+
+      await this.engineDelegate.afterCollect(result);
+
       if (commands.meta.description) {
         result.description = commands.meta.description;
       }
