+++ lib/firefox/webdriver/firefoxDelegate.js
@@ -9,6 +9,7 @@ const pathToFolder = require('../../support/pathToFolder');
 const rename = promisify(fs.rename);
 const { isAndroidConfigured, Android } = require('../../android');
 const geckoProfilerDefaults = require('../geckoProfilerDefaults');
+const util = require('../../support/util.js');
 
 class FirefoxDelegate {
   constructor(storageManager, options) {
@@ -209,6 +210,59 @@ class FirefoxDelegate {
       }
     } catch (e) {
       log.error('Could not get the HAR from Firefox', e);
+    }
+  }
+
+  async afterCollect(result) {
+    if (this.firefoxConfig.geckoProfiler && this.options.visualMetrics) {
+      for (let i = 0; i < result.length; i++) {
+        const index = i + 1;
+        const profileFilename = `geckoProfile-${index}.json.gz`;
+        const profileSubdir = pathToFolder(result[i].url, this.options);
+
+        try {
+          const geckoProfile = JSON.parse(
+            await this.storageManager.readData(profileFilename, profileSubdir)
+          );
+          if (!geckoProfile.meta) {
+            geckoProfile.meta = {};
+          }
+
+          // Here we are calculating the unix timestamp of the first frame after orange screen
+          // by adding the time to first frame, time to orange screen to the unix timestamp of
+          // video recording start time
+          const firstFrameStartTime =
+            result[i].recordingStartTime +
+            result[i].videoRecordingStart +
+            result[i].timeToFirstFrame;
+
+          for (let progress of [
+            'VisualProgress',
+            'ContentfulSpeedIndexProgress',
+            'PerceptualSpeedIndexProgress'
+          ]) {
+            result[i].visualMetrics[
+              progress
+            ] = util.adjustVisualProgressTimestamps(
+              result[i].visualMetrics[progress],
+              geckoProfile.meta.startTime,
+              firstFrameStartTime
+            );
+          }
+          geckoProfile.meta.visualMetrics = result[i].visualMetrics;
+
+          await this.storageManager.writeJson(
+            path.join(profileSubdir, `geckoProfile-${index}.json`),
+            geckoProfile,
+            true
+          );
+        } catch (e) {
+          log.error(
+            `Could not rewrite visual progress using startimes and add visual metrics to ${profileFilename}`,
+            e
+          );
+        }
+      }
     }
   }
