+++ lib/safari/webdriver/safariDelegate.js
@@ -19,6 +19,8 @@ class SafariDelegate {
 
   async beforeCollect() {}
 
+  async afterCollect() {}
+
   async onCollect(runner, index, result) {
     const resource = await runner.runScript(
       'return performance.getEntriesByType("resource");',
